+++ client/src/assets/scss/_variables.scss
@@ -1,0 +1,7 @@
+$gradient-background: linear-gradient(to right, #23315F, #101834);
+
+$color-darkest: #03050d;
+$color-input: rgba(255, 255, 255, 0.112);
+$color-text: #fff;
+
+$color-darkest-nav: rgba(3, 5, 13, 0.3);
