+++ client/src/components/LogoFull.vue
@@ -1,0 +1,11 @@
+<template>
+  <img src="../assets/logo_full.png" alt="Logo Full" >
+</template>
+
+<style lang="scss" scoped>
+  img {
+    max-width: 709px;
+    width: 60%;
+    min-width: 320px;
+  }
+</style>
