+++ client/src/pages/Profiles.vue
@@ -1,0 +1,3 @@
+<template>
+  <h1>Profiles</h1>
+</template>
