+++ client/src/pages/SignUp.vue
@@ -1,0 +1,3 @@
+<template>
+  <h1>Sign Up</h1>
+</template>
